Support for Artifact item slots. Contributed by hlwrave.

+++ L2J_Mobius_5.5_EtinasFate/java/com/l2jmobius/gameserver/model/conditions/ConditionUsingItemType.java
@@ -70,7 +70,7 @@ public final class ConditionUsingItemType extends Condition
 			
 			// So from here, chest armor matches conditions
 			
-			final int chestBodyPart = chest.getItem().getBodyPart();
+			final long chestBodyPart = chest.getItem().getBodyPart();
 			// return True if chest armor is a Full Armor
 			if (chestBodyPart == L2Item.SLOT_FULL_ARMOR)
 			{
