+++ L2J_Mobius_Classic_2.3_SevenSigns/java/com/l2jmobius/gameserver/model/items/L2WarehouseItem.java
@@ -171,7 +171,7 @@ public class L2WarehouseItem
 	/**
 	 * @return the part of body used with this item.
 	 */
-	public final int getBodyPart()
+	public final long getBodyPart()
 	{
 		return _item.getBodyPart();
 	}
